The files of the project were created and modified
index.html was created - inside of it there is an unordered list with
the links to the other pages of the site.
The folder "recipes" was created and inside of it 3 html pages for
each recipe.
Each page with a recipe has an image with the dish, a description of it,
an unordered list of ingredients and an ordered list of the steps to
cook it.
At the end of each recipe page there is a link with the source material.